+++ .gitignore
@@ -22,5 +22,8 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
-
 package-lock.json
+
+# dotenv environment variables file
+.env
+.env.test
