+++ api/.env
@@ -1,2 +1,8 @@
 Public Key = TEST-f30808a0-cdf7-4f27-96c9-4cd928c8ae34
 ACCESS_TOKEN = TEST-7006759729621021-081115-d84e26512c60341d764c895b757fb8ec-38250439
+DB_USER=postgres
+DB_PASSWORD=333333
+DB_HOST=localhost
+DB_NAME=starcards
+PORT=3001
+PGPORT=5432
