+++ client/index.html
@@ -51,9 +51,9 @@
                   <h3>Cart</h3>
                   <div class="summary-item"><span class="text">Subtotal</span><span class="price" id="cart-total"></span></div>
                   <!-- {Formulario para el envio de datos. Habria que colocar variables.} -->
-                  <form action="http://localhost:3000/checkout" method="post">  
+                  <form action="http://localhost:3001/mercadopago/checkout" method="post">  
                     <input type="hidden" name="title" value="Card StarCraft">
-                    <input type="hidden" name="price" value="2500">
+                    <input type="hidden" name="price" value="11">
                     <input type="submit" value="Buy Now" class="btn btn-primary btn-lg btn-block">
                   </form>
                   <!-- {Formulario para el envio de datos. Habria que colocar variables.} -->                </div>
